Refactor agentStore: Split slices, optimize API calls, and improve naming
