+++ src/features/ChatInput/ActionBar/Params/index.tsx
@@ -3,7 +3,7 @@ import { memo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { useAgentStore } from '@/store/agent';
-import { agentSelectors } from '@/store/agent/slices/chat';
+import { agentSelectors } from '@/store/agent/selectors';
 
 import Action from '../components/Action';
 import Controls from './Controls';
