+++ src/features/ChatInput/ActionBar/STT/browser.tsx
@@ -7,7 +7,7 @@ import { useTranslation } from 'react-i18next';
 import { SWRConfiguration } from 'swr';
 
 import { useAgentStore } from '@/store/agent';
-import { agentSelectors } from '@/store/agent/slices/chat';
+import { agentSelectors } from '@/store/agent/selectors';
 import { useChatStore } from '@/store/chat';
 import { operationSelectors } from '@/store/chat/selectors';
 import { useGlobalStore } from '@/store/global';
