+++ src/features/ChatInput/ActionBar/Search/FCSearchModel.tsx
@@ -5,7 +5,7 @@ import { Flexbox } from 'react-layout-kit';
 
 import InfoTooltip from '@/components/InfoTooltip';
 import { useAgentStore } from '@/store/agent';
-import { agentChatConfigSelectors } from '@/store/agent/slices/chat';
+import { agentChatConfigSelectors } from '@/store/agent/selectors';
 
 import FunctionCallingModelSelect from './FunctionCallingModelSelect';
