+++ src/features/ChatInput/ActionBar/Search/index.tsx
@@ -7,8 +7,7 @@ import { useTranslation } from 'react-i18next';
 import { useAgentEnableSearch } from '@/hooks/useAgentEnableSearch';
 import { useIsMobile } from '@/hooks/useIsMobile';
 import { useAgentStore } from '@/store/agent';
-import { agentSelectors } from '@/store/agent/selectors';
-import { agentChatConfigSelectors } from '@/store/agent/slices/chat';
+import { agentChatConfigSelectors, agentSelectors } from '@/store/agent/selectors';
 
 import Action from '../components/Action';
 import Controls from './Controls';
