+++ src/features/Conversation/Messages/components/Extras/TTS/index.tsx
@@ -2,7 +2,7 @@ import { memo, useMemo } from 'react';
 import { Md5 } from 'ts-md5';
 
 import { useAgentStore } from '@/store/agent';
-import { agentSelectors } from '@/store/agent/slices/chat';
+import { agentSelectors } from '@/store/agent/selectors';
 import { useGlobalStore } from '@/store/global';
 import { globalGeneralSelectors } from '@/store/global/selectors';
