+++ src/hooks/useModelHasContextWindowToken.ts
@@ -1,5 +1,5 @@
 import { useAgentStore } from '@/store/agent';
-import { agentSelectors } from '@/store/agent/slices/chat';
+import { agentSelectors } from '@/store/agent/selectors';
 import { aiModelSelectors, useAiInfraStore } from '@/store/aiInfra';
 
 export const useModelHasContextWindowToken = () => {
