+++ src/server/services/agent/type.ts
@@ -1,0 +1,6 @@
+import { LobeAgentConfig } from '@/types/agent';
+
+export interface UpdateAgentResult {
+  agent?: LobeAgentConfig;
+  success: boolean;
+}
