+++ src/store/agent/selectors.ts
@@ -1,1 +1,1 @@
-export { agentChatConfigSelectors, agentSelectors } from './slices/chat/selectors';
+export { agentChatConfigSelectors, agentSelectors } from './slices/agent';
