+++ src/store/agent/slices/agent/initialState.ts
@@ -4,7 +4,7 @@ import { DEFAULT_AGENT_CONFIG } from '@/const/settings';
 import { AgentSettingsInstance } from '@/features/AgentSetting';
 import { AgentItem, LobeAgentConfig } from '@/types/agent';
 
-export interface AgentState {
+export interface AgentSliceState {
   activeAgentId?: string;
   agentConfigInitMap: Record<string, boolean>;
   agentMap: Record<string, PartialDeep<AgentItem>>;
@@ -21,7 +21,7 @@ export interface AgentState {
   updateAgentMetaSignal?: AbortController;
 }
 
-export const initialAgentChatState: AgentState = {
+export const initialAgentSliceState: AgentSliceState = {
   agentConfigInitMap: {},
   agentMap: {},
   defaultAgentConfig: DEFAULT_AGENT_CONFIG,
