--- src/store/agent/slices/chat/index.ts
@@ -1,2 +1,0 @@
-export * from './action';
-export * from './selectors';
