--- src/store/agent/slices/chat/selectors/index.ts
@@ -1,2 +1,0 @@
-export { agentSelectors } from './agent';
-export { agentChatConfigSelectors } from './chatConfig';
