+++ src/store/agent/slices/knowledge/index.ts
@@ -1,0 +1,1 @@
+export { createKnowledgeSlice, type KnowledgeSliceAction } from './action';
