+++ src/store/agent/slices/plugin/index.ts
@@ -1,0 +1,1 @@
+export { createPluginSlice, type PluginSliceAction } from './action';
