+++ src/store/agent/store.ts
@@ -4,15 +4,23 @@ import { StateCreator } from 'zustand/vanilla';
 
 import { createDevtools } from '../middleware/createDevtools';
 import { AgentStoreState, initialState } from './initialState';
-import { AgentChatAction, createChatSlice } from './slices/chat/action';
+import { AgentSliceAction, createAgentSlice } from './slices/agent';
+import { KnowledgeSliceAction, createKnowledgeSlice } from './slices/knowledge';
+import { PluginSliceAction, createPluginSlice } from './slices/plugin';
 
 //  ===============  aggregate createStoreFn ============ //
 
-export interface AgentStore extends AgentChatAction, AgentStoreState {}
+export interface AgentStore
+  extends AgentSliceAction,
+    KnowledgeSliceAction,
+    PluginSliceAction,
+    AgentStoreState {}
 
 const createStore: StateCreator<AgentStore, [['zustand/devtools', never]]> = (...parameters) => ({
   ...initialState,
-  ...createChatSlice(...parameters),
+  ...createAgentSlice(...parameters),
+  ...createKnowledgeSlice(...parameters),
+  ...createPluginSlice(...parameters),
 });
 
 //  ===============  implement useStore ============ //
